backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
moviepy==1.0.3
numpy==1.24.3
pydub==0.25.1
python-magic==0.4.27
werkzeug==3.0.1
gunicorn==21.2.0
pytesseract==0.3.10
opencv-python==4.8.0.74
tensorflow==2.15.0
transformers==4.36.2
torch==2.1.2